"pages" of changeable content
Scroll to bottom of the test page, and I've added a new section. The default is a test to see how it handles long text, and below it is the option to go forward or back. I couldnt be bothered to type out that long text again so the first page is different after you star scrolling. turned off looping for now, would be helpful if you could possibly add animation later? For now, I've made it a set function with single purpose, but it can't be hard to make it access an array so it can have multiple different items which access it.