Changeable content
added ability to change content with "back/forawrds" buttons or text. Currently set to be used on the single part (at bottom of test page) to showcase, but should be easy to scale. Possibly by adding seperate arrays of text, and calling which one you want as part of the function? possibly a dictionary, but I'm not sure if that works in JS