variables hover animations
I just found out that CSS variables exist. The "close" button on the overlay now has the regular hover animation. now also possible to change the colour on the hover animations. I definitely got the "close" buttons animation wrong but I may or may not have deleted the original. Also more rounded overlay because why not.
Also, possibly 2 or more selectable themes for the website? We could put themes as a selectable option at the top, and put the main colours into certain variables, changed with that button click? just a thought, not really necessary.